refactor(core): move Pool unit tests into pool.rs
Session-pool checkout/checkin/close/FIFO tests live next to Pool now.
engine.rs keeps Engine::split_pool (it wraps sizing + Pool::new).

Constraint: no production behavior change
Confidence: high
Scope-risk: narrow
